+++ experimental/packages/opentelemetry-sdk-node/src/start.ts
@@ -19,15 +19,12 @@ import {
   getIdGeneratorFromConfiguration,
   getSamplerFromConfiguration,
   getInstanceID,
-  getMeterReadersFromConfiguration,
-  getMeterViewsFromConfiguration,
   getResourceDetectorsFromConfiguration,
   getResourceFromConfiguration,
   getSpanProcessorsFromConfiguration,
 } from './utils';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import type { SDKComponents, SDKOptions } from './types';
-import { MeterProvider } from '@opentelemetry/sdk-metrics';
 import { TracerProvider } from '@opentelemetry/sdk-trace';
 import { logs } from '@opentelemetry/api-logs';
 import type {
@@ -45,6 +42,7 @@ import { ATTR_SERVICE_INSTANCE_ID } from './semconv';
 import { diagLogLevelFromSeverityNumberConfig } from './diag';
 import {
   createLoggerProviderFromConfig,
+  createMeterProviderFromConfig,
   createPropagatorFromConfig,
   createSpanLimitsFromConfig,
 } from './create-from-config';
@@ -90,7 +88,9 @@ export function startNodeSDK(sdkOptions?: SDKOptions): {
   try {
     components = create(config, sdkOptions);
   } catch (createErr) {
-    diag.error(`Could not create OpenTelemetry SDK: ${createErr.message}`);
+    diag.error(
+      `Could not create OpenTelemetry SDK from configuration, SDK will not be setup: ${createErr.message}`
+    );
     return NOOP_SDK;
   }
   if (components.contextManager) {
@@ -156,15 +156,11 @@ function create(
       );
     }
 
-    const meterReaders = getMeterReadersFromConfiguration(config);
-    if (meterReaders) {
-      const meterViews = getMeterViewsFromConfiguration(config);
-      const meterProvider = new MeterProvider({
-        resource: resource,
-        readers: meterReaders,
-        views: meterViews ?? [],
-      });
-      components.meterProvider = meterProvider;
+    if (config.meter_provider) {
+      components.meterProvider = createMeterProviderFromConfig(
+        resource,
+        config.meter_provider
+      );
     }
 
     const spanProcessors = getSpanProcessorsFromConfiguration(config);